pyproject.toml
[project]
name = "fastapi-celery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "fastapi[standard]>=0.116.1",
    "pydantic-settings>=2.10.1",
    "redis>=6.4.0",
]
